Split the PrivacyExperience endpoints and PrivacyExperienceConfig endpoints into separate files for easier review.
