+++ src/fides/api/ops/api/v1/api.py
@@ -13,6 +13,7 @@ from fides.api.ops.api.v1.endpoints import (
     oauth_endpoints,
     policy_endpoints,
     policy_webhook_endpoints,
+    privacy_experience_config_endpoints,
     privacy_experience_endpoints,
     privacy_notice_endpoints,
     privacy_preference_endpoints,
@@ -38,6 +39,7 @@ api_router.include_router(oauth_endpoints.router)
 api_router.include_router(policy_endpoints.router)
 api_router.include_router(policy_webhook_endpoints.router)
 api_router.include_router(privacy_experience_endpoints.router)
+api_router.include_router(privacy_experience_config_endpoints.router)
 api_router.include_router(privacy_notice_endpoints.router)
 api_router.include_router(privacy_preference_endpoints.router)
 api_router.include_router(privacy_request_endpoints.router)
